move pyg_data_transform to files to be removed
(cherry picked from commit 28d4bf08e4f96267f23d937ff9f415d336247d6f)
