+++ src/fairchem/core/datasets/oc22_lmdb_dataset.py
@@ -16,9 +16,9 @@ import torch
 
 from fairchem.core.common.registry import registry
 from fairchem.core.common.typing import assert_is_instance as aii
-from fairchem.core.common.utils import pyg2_data_transform
 from fairchem.core.datasets._utils import rename_data_object_keys
 from fairchem.core.datasets.base_dataset import BaseDataset
+from fairchem.core.datasets.lmdb_dataset import pyg2_data_transform
 from fairchem.core.modules.transforms import DataTransforms
